InnoDB memcached interface requirement: Transfer callback array to memcached
InnoDB API memcached interface engine expects InnoDB API callback
array pointer as a 'data' member of its st_plugin_int.

This is how it is implemented in MySQL, but a better solution
should be figured out. The requirement of this solution is to have
memcached plugin initialize after the InnoDB plugin, so the logic
has been changed to initialize it as the last one plugin.

NOTE: This commit can be reverted when InnoDB API memcached interface
      will be deleted (e.g. when it will be replaced by Handler API
      memcached interface).