Assertions are disabled on release builds in MariaDB
MariaDB uses -DNDEBUG flag when compiling release builds (MySQL does
not do that). Because of that, assertions are disabled, what means
that ib_cb_trx_start() is not called here on release builds.

This results in crash during handling of 'flush_all' memcached
command, because transaction is not started properly.