Fix multiple memcached_sync_count for the same transaction
The memcached_sync_count is incremented multiple times for the same
transaction. It is incremented each time innodb_conn_init() is called,
i.e. on the beginning of processing of every memcached command.

However, the memcached_sync_count is decremented only on transaction
commit/rollback, in innodb_reset_conn(). This means, that after the end
of transaction, it is not reset to zero.

This blocks TRUNCATE command on any InnoDB table, on which memcached
has operated during program lifetime. Even after uninstalling the
plugin.

This fix ensures that ib_cb_cursor_set_memcached_sync() is called
only once for each transaction.